+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Fixed
 
 - Give an error message when the user's filters (by level, by topic, by concept identifier) don't match any concepts. Fixes [#437](https://github.com/fniessink/toisto/issues/437).
+- Give an error message when the target and source language are the same. Fixes [#453](https://github.com/fniessink/toisto/issues/453).
 
 ### Added